Move `shape_util.hpp` to `dev_api/openvino/core/`
