+++ src/core/src/op/mod.cpp
@@ -6,8 +6,8 @@
 
 #include "element_visitor.hpp"
 #include "itt.hpp"
+#include "openvino/core/shape_util.hpp"
 #include "openvino/reference/mod.hpp"
-#include "shape_util.hpp"
 
 namespace ov {
 namespace op {
