+++ src/core/src/op/util/variable_value.cpp
@@ -9,10 +9,10 @@
 #include "ngraph/runtime/host_tensor.hpp"
 #include "openvino/core/deprecated.hpp"
 #include "openvino/core/shape.hpp"
+#include "openvino/core/shape_util.hpp"
 #include "openvino/runtime/allocator.hpp"
 #include "openvino/runtime/itensor.hpp"
 #include "openvino/runtime/tensor.hpp"
-#include "shape_util.hpp"
 
 OPENVINO_SUPPRESS_DEPRECATED_START
 namespace {
