+++ src/core/src/runtime/itensor.cpp
@@ -7,10 +7,10 @@
 #include <memory>
 
 #include "openvino/core/except.hpp"
+#include "openvino/core/shape_util.hpp"
 #include "openvino/runtime/allocator.hpp"
 #include "openvino/runtime/iremote_tensor.hpp"
 #include "openvino/runtime/properties.hpp"
-#include "shape_util.hpp"
 
 namespace ov {
